Various improvements from code review.
